// srcc/app/shared/models/customer.model.ts
import Address from './address.model';

export default class Customer {
    selected: boolean = false;
    id: string = null;
    first_name: string = null;
    last_name: string = null;
    company_name: string = null;
    email: string = null;
    phone: string = null;
    bank_account: string = null;
    address: Address = null;
    label: string = null;
    name_label: string = null;
}